apex-enrichment-agent/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
dnspython==2.7.0
python-dotenv==1.0.1
supabase==2.9.1
anthropic==0.39.0
tldextract==5.1.3
tenacity==9.0.0
